moved password verification to request handler
